travis: run unit tests on MacOS
This requires a little bit of working around MacOS-specific weirdnesses
(which are fine under POSIX but don't match Linux). The most obvious one
is the lack of 'readlink -f', but there's also the lack of sub-second
granularity of filesystem timesamps, and the default-follow behaviour of
link(2) when trying to hardlink a symlink.

Signed-off-by: Aleksa Sarai <asarai@suse.de>