test: correctly collate unit test coverage
Because -coverprofile truncates the target file, we were ignoring the
rootful unit test coverage (resulting in the MakeRead{Only,Write} calls
being counted as "untested"). Luckily most of the code was being
exercised by the integration tests (which did do the collation
correctly), but that means this only slightly increases our overall
coverage.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>