cmd: move version to github.com/opencontainers/umoci
Because of how '-ldflags -X' works, if we keep the version string in the
main package we have to have different build flags when doing "go test"
since in test-mode the main package is no longer ".../cmd/umoci".

In future we should probably get rid of VERSION and just store the
version inside the Go source code so that "go get" gives you a valid
version.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>