site: switch to hosting on netlify
The main reasons for this move are that we can explicitly set our own
HTTP security headers (some of which are required for the gold CII
badge). In addition, the TLS certificate used is per-domain and is
managed by Netlify (through Let's Encrypt) instead of CloudFlare.

In addition this ensures that all aspects of the umo.ci website can be
managed by the OCI.

Signed-off-by: Aleksa Sarai <asarai@suse.de>