GC: add support for GC policies
For our dist-spec implementation [1][2],
we have a use case where we use umoci's GC code to clean up orphaned
blobs. The issue is that as per dist-spec layer upload and manifest
updates are two different API calls and typically in that order, which
means layers begin life as orphans.

Adding the proposed patch allows us to:
a) mitigate the above issue
b) have a generic policy framework for future expansion

[1] https://github.com/opencontainers/distribution-spec
[2] https://github.com/anuvu/zot

Signed-off-by: Ramkumar Chinchani <rchincha@cisco.com>