oci: casext: clean up Reachable and drop Paths
Reachable didn't make use of an obvious optimisation (don't walk into
digests which we've seen before), and there are no users of Paths at the
moment so we can safely drop it for now.

In addition, Reachable is only used by GC -- and its semantics only make
sense for GC -- so un-export it accordingly.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>