layer: add support for overlayfs style whiteouts
While not strictly OCI standard, it can be useful to have umoci write
overlayfs stile whiteouts when unpacking instead of just manipulating the
filesystem on its own, for example for use with drivers which are going to
use overlayfs to speed up image building and extraction.

Note that writing whiteouts (i.e. mknod foo c 0 0) is supported as an
unprivileged user in the upstream kernel via a3c751a50fe6 ("vfs: allow
unprivileged whiteout creation"), which will be released in 5.8. There is
*no* support right now for writing the opaque overlay attribute as an
unprivileged user, but since that's mostly done via GenerateInsertLayer()
and not in the general case, I doubt most people will run into this. And if
they do, they can send a kernel patch :)

Note that we only run the test on linux, since osx presumably doesn't have
this hack to let us mknod c 0 0 as an unprivileged user.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>