mutate: implement an AddExisting() API
in order to add layers that may already exist in the image to a mutator, we
would previously have to read the layer off the disk, decompress it, and
recompress it (arriving at the same results, modulo gzip stream weirdness).

Let's instead just implement an API for letting us do this directly,
without all the i/o and CPU usage.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>