oci: cas: add StatBlob to CAS interface
With VerifiedReadCloser, users are in a situation where using GetBlob
for pure existence checks results in the entire blob being read from
disk on Close() even though they didn't read any bytes (meaning the
verification check isn't protecting against anything).

We can't make VerifiedReadCloser a no-op in the read-nothing case, but
it also seems that we should have a separate API for this anyway --
hence a stat-ish API that just does a "does this exist" check.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>