oci: config: correctly handle runtime-spec versions
Hardcoding "1.0.0" as a version for all config.json output is not
correct, but we need to detect unsupported original config versions.

Since we cannot be sure what a future version of the runtime-spec might
add, we can only support runtime-spec versions that have the same major
version number and are not newer than the runtime-spec version we have.
This should be conservative enough to avoid most issues.

With that in place, we can output our own runtime-spec version to all
config.json files we generate (ensuring we're more spec-compliant).

(I'm not sure how widely this function is used outside of umoci.)

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>