oci: config: strip "-dev" from runtime-spec version
Unfortunately, the current way runtime-spec uses SemVer is technically
incorrect since using <version>-dev indicates a pre-release of
<version>, not an unreleased branch after the release of <version>.
Arguably SemVer doesn't quite support this feature, but using "+dev" (as
runc and umoci do) would avoid this problem.

In addition, runtime-tools requires the runtime-spec version to match
the version it was compiled with exactly in order to be able to do any
validation. Arguably this is kind of unnecessary but it is currently
designed to be overly-conservative. However, we sometimes need to use
non-released versions of the spec (which is still compatible with the
released version) and doing so causes issues with validation.

It also is better to not include "-dev" in versions we output, purely
for the reason that it results in more reasonable-looking
configurations.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>