test: move 'go test' output to stderr for coverage binary
In Go 1.20 and newer, it seems that the coverage output was added to the
test output (which is printed to stdout by default), breaking our tests.

The nicest long-term solution would be to switch to "go build -cover"
which was added in Go 1.20, and remove our main_test.go logic, but this
patch is a simpler stop-gap to fix the test failures.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>