fix: handle overlay xattr opaque bit
Current behavior determines if a path is a whiteout if a overlay char
dev is present.

Additionally, also check the extended attrs.

For files, this is not important: if a file /foo/bar exists in
some two layers, then the higher layer's version will replace
the lower layer's version whether or not the opaque bit is set.

However, directories behave differently.  If /foo exists in two
layers and no opaque bit is set, then we see the union of their
contents.  But if the opaque bit is set on the higher layer's
version, then we should not see the lower layer's version.

So, in that case, create a whiteout /.wh_foo.

Signed-off-by: Ramkumar Chinchani <rchincha@cisco.com>