*: migrate to Go stdlib error wrapping
Quite a few of these callers don't need to use fmtcompat, but for
simplicity they have all been converted and we can slowly remove
fmtcompat.Errorf calls when we verify that they are correct.

Some of the obvious "err = nil" logic has been reworked, but there are
still some suble usages of it that will take some time to find and fix.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>