mutate: make uncompressed annotation hint not use globals
The previous implementation of the BytesRead() would store the value of
bytes read in the compressor implementation singleton, meaning that if
you were to compress multiple data streams you would end up with a
meaningless value from BytesRead(). Not good.

It also makes little sense to have the byte-counting logic embedded in
the compression code -- there is only one user that cares about this and
it's trivial to add a dummy io.Reader wrapper that counts the bytes
read.

Fixes: 8f65e8f75478 ("mutate: add uncompressed blob size annotation")
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>