unpack: make layer decompression based on mediatype more generic
Rather than hard-coding a set of media-types, parse the "+foo" MIME type
suffix to figure out what compression algorithm was used for the layer.

This also enables zstd decompression support out of the box (since that
is a registered algorithm now) and also improves our error messages --
we now provide a more helpful error to explain whether the issue is that
the underlying MIME type of the layer is unsupported (think squashfs) or
if it's just an issue with the compression algorithm.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>