repack: don't hardcode layer compression algorithm
If a user doesn't specify what compression type they would like to
(*Mutator).Add(), auto-select the compression algorithm by first trying
to use whatever the last layer's compression algorithm was (if we
support it) followed by a fallback to a default (gzip for now).

It is preferable to try to re-use the previous compression algorithm so
that users which operate on zstd-compressed images will produce
zstd-compressed layers rather than going with gzip.

For now, make this the default for the CLI as well. A future patch will
add support for a --compress=... flag that will let you configure this
behaviour and explicitly set a compression method.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>