pkg: mtreefilter: fix broken err check in tests
It turns out that I actually did need glasses back in 2017, because
commit d027fb542757 ("pkg: add mtreefilter package") incorrectly used
"if err != ..." rather than "if err := ...".

Luckily the tests were still correct (err must be nil at this point in
the tests, so this was a complicated way of writing "if nil != ...") but
obviously this wasn't what was intended.

This was the only instance of "if err != " followed by something not
"nil" in our codebase.

Fixes: d027fb542757 ("pkg: add mtreefilter package")
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>