oci: generate: support converting trusted.overlay.{whiteout,opaque} xattrs
As these are officially support ways of creating whiteouts with
overlayfs, we should support converting them. While the
trusted.overlay.whiteout xattr is probably very rare to see in the wild,
trusted.overlay.opaque=y is probably something we are going to encounter
and so we should support it.

It's not clear what we should do about trusted.overlay.opaque=x -- the
documentation (and code) imply that it has similar behaviour to
trusted.overlay.opaque=y but in practice directories marked with that
xattr don't actuall act like an opaque directory. We'll just ignore it
for now.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>