oci: extract: correctly handle mixed overlayfs whiteouts
As per the overlayfs documentation[1], extracting plain overlayfs
whiteouts inside an opaque whiteout directory is not recommended:

> A directory is made opaque by setting the xattr
> "trusted.overlay.opaque" to "y". Where the upper filesystem contains
> an opaque directory, any directory in the lower filesystem with the
> same name is ignored.
>
> An opaque directory should not contain any whiteouts, because they do
> not serve any purpose.

In practice, it results in unfortunate readdir behaviour where the plain
whiteouts are listed by readdir despite not being accessible. Aside from
potention information leaks (proving that lower layers contain a file
that has been masked) it also can lead to user confusion.

Unfortunately the solution for this is kind of complicated, as in order
to support the OCI requirements for opaque directories we need to keep
any existing paths we have unpacked but we need to exclude whiteout
inodes from being created inside opaque whiteout directories.

The simplest solution is to have a trie (similar to upperPaths) which
tracks the whiteouts we have extracted, and we can then deleting
existing whiteouts or avoid extracting whiteouts based on the state of
whiteouts in the upper layer. While this could've been merged with
upperPaths, upperPaths contains far more entries and is stored more
efficiently than upperWhiteouts -- for whiteout handling in overlayfs
it's more efficient for us to only walk the subset of whiteout entries
(which should be very small for most images).

[1]: https://www.kernel.org/doc/html/v6.14/filesystems/overlayfs.html#whiteouts-and-opaque-directories

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>