oci: extract: properly handle missing parent dirs for overlayfs
The main thing to consider here is that when extracting stuff for
overlayfs we are dealing with one layer in the layer stack, which means
that we may be creating whiteouts for paths that don't exist or have the
wrong type.

In practice most archives will contain parent directories, but it's not
a given and we should be able to handle these kinds of edge cases (like
we do in the regular extraction mode).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>