pkg: pathtrie: optimise deletions to prune more garbage
The previous logic would only prune unused branches of the trie in very
specific (and trivial) circumstances, but it is quite easy for us to
optimise the pruning by simply tracking how many children there are
all in all descendants of the trie.

This lets us quickly decide whether a branch is dead and should be
pruned.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>