oci: layer: merge is-overlayfs config and mapping into OnDiskFormat config
Previously we had separate ways of specifying the on-disk format with
UnpackOptions and RepackOptions (UnpackOptions had an enum for the
on-disk format while RepackOptions just had a boolean specifying whether
overlayfs-style whiteouts should be translated).

Conceptually it makes far more sense to think of the configuration as
being an "on-disk format", so we should represent it that way in the
configuration. In addition to the on-disk format type, the userns
mapping applied to the rootfs is more a property of the on-disk format
than being a standalone option.

By creating an OnDiskFormat interface that has multiple implementations
we can also add per-on-disk-format configuration options (such as
userxattr for the overlayfs on-disk-format, which will be added later in
this series).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>