oci: layer: pass RepackOptions to tarGenerator
This mirrors TarExtractor, and lets us provide information about
overlayfs whiteouts to the generator.

Note that this is a bit of a temporary fix -- ideally we would unify the
on-disk format configuration between TarExtractor and tarGenerator.
Unfortunately this is a little complicated because WhiteoutMode was
added to umoci.json back in commit 9ab5eff081e2 ("meta: include the
WhiteoutMode") so it will require some compatibility considerations to
fully implement in a later patchset.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>