oci: layer: refix auto-applied xattr handling
We've had special handling of xattrs like security.selinux (which the
system auto-applies when inodes are created) for a very long time, but
commit 9a1cefaf16bf ("oci: layer: correctly handle trusted.overlay xattr
namespace escaping") accidentally broke this during a refactor.

The solution is to simply invert the filter.MaskedOnDisk check, but we
need to have some tests that simulate security.selinux to ensure we
don't make the same mistake again in the future.

Fixes: 9a1cefaf16bf ("oci: layer: correctly handle trusted.overlay xattr namespace escaping")
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>