pkg: hardening: fix spurrious InvalidDigest test failure
For tests with a buffer size of 1, if the byte was equal to our "fake"
byte (0x80) then the test would fail because the digest would be valid.
Since it's a single byte, the chance of it happening was 1-in-256, but
obviously with enough CI runs it happens every once in a while.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>