oci: blobcompress: switch to Docker-friendly gzip block size
According to the stacker folks, having a parameter that mismatches with
Docker will result in layers getting recompressed. Based on my testing,
the hashes of random layer data doesn't change but I guess it's possible
it could change in certain situations and we should just match their
settings.

Ideally this would be configurable but since stacker is the only thing
that cares about this for the moment, we may as well just follow their
default setting (which -- it turns out -- is the new pgzip default
setting too).

Suggested-by: Ramkumar Chinchani <rchincha@cisco.com>
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>