lint/staticcheck: apply minor fixups
Most of the fixes (ST1006, S1021, QF1004) are fairly uncontraversial.

However for QF1001, I opted to ignore all of the proposals because the
"more complicated" version is actually easier to understand if you want
to only allow something if one specific condition is not true. Just
compare

  if !(a && b && c) {
      // do something if all of (a, b, c) are not true
  }

to

  if !a || !b || !c {
      // do something if any of (a, b, c) are false
  }

For complicated boolean statements the former is sometimes easier to
understand, which happens to be the case here for us.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>