lint/forcetypeassert: make type assertion requirements more standard
The main user of type assertions in umoci is the cli metadata handling
for passing around pre-parsed arguments. Now that Go has generics we can
create a generic helper to explicitly assert that the types match with a
helpful panic message (incorrectly expecting the wrong type or a
requirement metadata argument missing is a programming error, hence the
panic).

The only other two users were already safe as the type was checked in a
less direct way (in the case of mapDescriptors) or the type is
definitely known to be correct (in the case of our CAS tests).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>