casext: fix typo in t.Run
During the rework in commit 1c93e3f0f49a ("tests: modernise and switch
to testify assertions"), we accidentally removed the variable name for
the *testing.T passed to the t.Run closure in a single test, which would
make test failures fail in somewhat unexpected ways.

A quick grep shows that this is the only instance of this happening in
the codebase.

Fixes: 1c93e3f0f49a ("tests: modernise and switch to testify assertions")
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>