tests: switch to modern GOCOVERDIR
This removes the need for us to have hacky collation of coverage
reports, and lets us just use a single directory for all test runs. For
our actual CI we need to merge the GOCOVERDIR from different runs, but
this is fairly trivial with "go tool covdata".

The main benefit is that this removes the "go test -c" and TestUmoci
hack that was a common cause of fun issues for tests that checked the
output of umoci commands.

Unfortunately, there appears to be a Go compiler bug in "go test" that
we need to work around in our unit test wrapper.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>