oci: dir: ensure ownership of new files matches image dir ownership
"sudo umoci repack" when done on non-root-owned image layout directories
results in new files having file ownerships that are quite problematic.

We could special-case this for sudo by looking at SUDO_[UG]ID but it
seems more appropriate to make sure that we match the owner of the
layout directory. The owner of the layout directory can practically
create and delete any subpaths so they are conceptually the right user
to emulate in this case.

From a security perspective, the worst that can happen is that the image
owner can access the new blobs we write, but surely that is something
the user intended in the first place.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>