dockerfile: fix image-tools build
Go 1.22 removed support for auto-converting old-school vendoring formats
to go.mod, which meant that we implicitly built with the newer spec
versions (this should have failed CI for other reasons, but it seems we
didn't see it).

image-spec has recently defined a new module in ./schema, which results
in failures when trying to auto-pull the dependencies needed to build
image-tools.

Eventually we need to get rid of all of this and implement verification
in umoci...

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>