umoci: add raw-mtree-validate helper to replace "gomtree --rootless"
The go-mtree CLI does not support rootless operations, which requires us
to use a patched version maintained in my OBS home project. This is a
little silly, and really we should just have a helper that calls
mtree.Check() with fseval.Rootless when in rootless mode.

At the moment, we only support the minimal set of flags needed for our
tests and the command is hidden from our help pages.

Note that we need the (not available in gomtree) "--umoci-keywords" flag
for a similar reason to the one outlined in commit 24e77f0f6e3f ("cmd:
hardcode go-mtree keywords") -- in certain contexts we could end up with
a manifest that has fewer keywords than the set we want and so using it
as the source of truth is somewhat suspect and can lead to unfortunate
bugs.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>