umoci: add raw-mtree-validate helper to replace "gomtree --rootless"
The go-mtree CLI does not support rootless operations, which requires us
to use a patched version maintained in my OBS home project. This is a
little silly, and really we should just have a helper that calls
mtree.Check() with fseval.Rootless when in rootless mode.

At the moment, we only support the minimal set of flags needed for our
tests and the command is hidden from our help pages.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>