test: use "umoci raw mtree-validate" in tests
While it would be nice to remove gomtree entirely from our tests, I am a
little apprehensive about having all of our validation logic for umoci
also be inside umoci. Instead, we use both in privileged mode and return
an error if the status is different between the two.

With this we can finally get rid of the awful patched gomtree binary
maintained in my OBS home project.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>