tests: use --strict mode for gomtree validation
The default setting for what kinds of manifest validation failures will
cause "gomtree validate" to return an error are really suboptimal for
our usecase, as they are designed for compatibility with FreeBSD. In
particular, only modified files where a keyword was modified (as
opposed to being missing) would result in errors.

The new --strict flag allows us to get something more like what we want
(any discrepancy between the manifest and the rootfs gives us an error).
This is broadly in line with what umoci already does (both during diff
layer generation and "umoci raw mtree-validate").

In practice, nothing changes because we didn't have any bugs in this
area, but at least we can now be sure that we will detect bugs here.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>