deps: bump to image-spec v1.1.1
There were a few build failures related to the embedding of the
ispec.Platform struct, as well as some lint errors due some fields being
marked as deprecated.

In addition, add some new snapshot tests for pretty-printing the new
examples in the spec. As we do not yet support pretty-printing this
data, these tests will be adjusted in a future patch in this series when
we add support for it.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>