mutate: handle DescriptorPath rewriting more delicately
When rewriting the DescriptorPath, we would previously only modify
descriptors that exactly match our old descriptor and replace all of the
fields.

This is less than ideal -- in principle we want to replace all
descriptors that have the same digest, but without clearing any fields
(to avoid clearing annotations from descriptors that have extra
annotations which is conceivable).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>