stat: do not error out for "invalid" config blobs
With the existence of OCI arifacts we should at least provide some
information with "umoci stat" rather than erroring out if we see an
image that has a configuration blob that is non-standard. Other
operations will still fail, of course.

We will still error out if the manifest is invalid -- this is mainly
because the stat code already implicitly requires the top blob to be a
manifest (and our reference resolution logic only targets manifests
currently). Also, even OCI artifacts still use manifests (they were
forced to because of similar issues with registries).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>