config: add "platform" prefix to "os" and "architecture" setting names
This applies to both the Go API (breaking) and "umoci config" CLI (with
backward-compatible aliases). The main reason is that now that v1.1 of
the image-spec decided to properly refactor the platform configuration
into a separate structure, it makes sense for us to refer to it that way
in our own APIs.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>