tests: add handling for non-test binaries
openSUSE QA has started running our tests with the actual package
binaries, which causes the xattr tests to fail (because we assume that
the binary is a test binary with special handling for forbidden xattrs).

The solution is to provide some information in "umoci --version" about
whether the binary is a test binary and then modify the test
accordingly.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>